Stop crashing Smugleaf's game
The assembly being generated by the JIT compiler doesn't have actual
access to widgets inside Robust.Client. It probably should.

There's a companion change to our fork of XamlX that gives XamlX access
to types internal to Robust.Client.

Note that this change is only needed if you've hacked XamlX to be able
to see types inside of Robust.Client.