Fix replays playing more sound effects than they should.
Replays in SS14 play a lot of sounds globally they shouldn't. For example, antag greeting sounds. This makes for a very annoying viewing experience.

The cause of the issue is that these sounds are played with PlayGlobal(), but with a player filter. In normal play, this player filter causes the game to *not* send audio data to players it shouldn't. Replays, however, always receive these audio entities, so they play the sound anyways.

The audio system already recorded which entities should be able to hear sounds (based on their player actors). We now include this data in replays via a new AudioPlayerFilter component.

This alone is not enough, however. This allows us to accurately handle filtering sounds when spectating active players, but it is also possible in replays to spectate random entities (or fly around with an observer that isn't in the replay at all). Simply blocking all sounds that have a filter would cut out many crucial global sounds.

To solve this problem, I added a FilterExpression type that is automatically tracked by Filter. It's a best-effort approximation of the filter, that can be evaluated client-side. The only case covered right now is that it allows distinguishing global broadcast filters and player-specific filters, so that we can properly handle these sounds.

Evaluation of whether we even *should* use the FilterExpression is done based on a new ReplayCamera component, which content will need to set in its replay playback logic.