feat: duplicate localization keys for downstreams
Previously the code collected occurrences of duplicate localization keys
 and generated log errors which induced test failure. This is good for
discovering unintended duplicates, but downstreams sometimes intend to
override localizations. This change checks if the resource directory
begins with an underscore (as is custom in downstream codebases) and
does not attempt to emit log errors for duplicate keys. Consequently
they will no longer be informed of duplicated keys even outside of the
main namespace. Perhaps this is an acceptable tradeoff.