Prevent concurrent modification of InputManager.UIKeyBindStateChanged
Changing devwindow to use the editor dockers seems to change the exact behavior of the UI tab in such a way that it causes a concurrent modification of this event...

Make it an immutable list. This is an extremely low-level event so the allocation overhead on change is not a big deal here.